requirements.txt
openai
python-dotenv
langchain-community
duckduckgo-search
streamlit